Breaking down functionalities into Components - ready to start Calendar integration Function
